Finding most similar artist.
I return the most similar artist in a tuple of (name, id) so we can use this later on.  I need to write down that difflib and numpy are included down in the readme.  I also need to document this portion of the code.  It would be good to further generalize this in case it's presented in the form of a list instead of a dict.