adjacency matrix formed
Here, I form the adjacency matrix.  The difference between this and the original build is that the original worked under the top of songs of the artist.  The current build gets them for all songs of an artist, hence this one is going to be sparse for many nodes but dense for a few critical points.